backend/requirements.txt
fastapi==0.112.2
uvicorn>=0.15.0,<0.16.0
pydantic>=2.0.0
tortoise-orm
asyncpg
sqlalchemy
great_expectations